requirements.txt
django-cors-headers==3.13.0
Django==4.1
djangorestframework-camel-case==1.3.0
djangorestframework==3.13
drf_yasg==1.20.0
numpy==1.23.5
pandas==1.5.2
psycopg2==2.9.*
pytest-django==4.5.2
pytest-env==0.8.1
pytest-html==4.0.2
python-dotenv==0.21.*
pylint-django==2.5.3
pytest-cov==4.1.0
pytest-order==1.1.0
python-dateutil==2.8.2